Implement task2. Add iUnit interface. Implement figts between army. Implement unittests

+++ SoftGame/Configuration/UnitsCoufiguration.cs
@@ -2,10 +2,10 @@
 {
     public static class UnitsCoufiguration
     {
-        public static double WarriorHealth = 50;
+        public static int WarriorHealth = 50;
         public static double WarriorAttack = 5;
         
-        public static double KnightHealth = 50;
+        public static int KnightHealth = 50;
         public static double KnightAttack = 7;
     }
 }
